requirements.txt
Django>=5.0.0,<5.1.0
Pillow>=10.0.0,<11.0.0
django-crispy-forms>=2.0.0,<3.0.0
crispy-bootstrap5>=0.7,<1.0
faker>=18.0.0,<19.0.0
python-dotenv>=1.0.0,<2.0.0
whitenoise>=6.5.0,<7.0.0
gunicorn>=21.0.0,<22.0.0
dj-database-url>=2.0.0,<3.0.0
psycopg2-binary>=2.9.0,<3.0.0
